checks for db dir before deleting it.

+++ lib/src/hive_storage_service_base.dart
@@ -38,6 +38,8 @@ class HiveStorageService {
   Future wipe() async {
     var appDir = await getApplicationDocumentsDirectory();
     var hiveDb = Directory('${appDir.path}/$subDirectory');
-    hiveDb.delete(recursive: true);
+    if (await hiveDb.exists()) {
+      hiveDb.delete(recursive: true);
+    }
   }
 }
